refactor: 커스텀 select를 html 기본 select 태그로 변경
