+++ src/feature/CardRegister/components/InfoInputSection/CardCompanySelectField/CardSelectField.tsx
@@ -1,9 +1,21 @@
 import { useState } from "react";
-import CardCompanyOptionList from "./CardCompanyOptionList/CardCompanyOptionList";
 import type { CardCompanyType } from "../../../../../common/types/CardCompany";
+import styled from "styled-components";
+import { colors } from "../../../../../styles/color";
+import CardCompanyOption from "./CardCompanyOption/CardCompanyOption";
+
+const mockCardCompany: CardCompanyType[] = [
+  "BC카드",
+  "신한카드",
+  "카카오뱅크",
+  "현대카드",
+  "우리카드",
+  "롯데카드",
+  "하나카드",
+  "국민카드",
+];
 
 const CardCompanySelectField = () => {
-  const [isOpen, setIsOpen] = useState(true);
   const [selectedCardCompany, setSelectedCardCompany] = useState("");
 
   const handleCardCompanyClick = (cardCompany: CardCompanyType) => {
@@ -12,22 +24,35 @@ const CardCompanySelectField = () => {
     }
   };
 
-  const handleCardSelectClick = () => {
-    setIsOpen((previous) => !previous);
-  };
   return (
-    <div>
-      <button type="button" onClick={handleCardSelectClick}>
-        카드사를 선택해주세요
-      </button>
+    <CardSelectBox>
+      <CardSelectTrigger
+        defaultValue={""}
+        onChange={(e) =>
+          handleCardCompanyClick(e.target.value as CardCompanyType)
+        }
+      >
+        <option value={""} disabled hidden>
+          카드사를 선택해주세요
+        </option>
 
-      {isOpen && (
-        <CardCompanyOptionList
-          handleCardCompanyClick={handleCardCompanyClick}
-        />
-      )}
-    </div>
+        {mockCardCompany.map((cardCompany) => (
+          <CardCompanyOption cardCompany={cardCompany} key={cardCompany} />
+        ))}
+      </CardSelectTrigger>
+    </CardSelectBox>
   );
 };
 
 export default CardCompanySelectField;
+
+const CardSelectBox = styled.div`
+  display: flex;
+  position: relative;
+`;
+
+const CardSelectTrigger = styled.select`
+  width: 100%;
+  border: 1px solid ${colors.border.default};
+  text-align: left;
+`;
